backend/requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
colorama==0.4.4
decorator==5.1.1
Django==3.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
dnspython==2.1.0
drf-extra-fields==3.4.0
idna==3.3
ipython==7.16.2
jedi==0.17.2
matplotlib-inline==0.1.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.29
Pygments==2.12.0
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.2.4
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
